Fix failed tests
