+++ django_project/cplus_api/tasks/__init__.py
@@ -2,4 +2,4 @@ from .cleaner import *  # noqa
 from .runner import *  # noqa
 from .remove_layers import *  # noqa
 from .verify_input_layer import *  # noqa
-from .sync_default_layers import *
+from sync_default_layers import sync_default_layers
